Conflicts:
	src/Test/EnnemisBasiquesEtSpecialesTest.java
	src/TowerDefense/controleur/Controleur.java
	src/TowerDefense/modele/Jeu.java
	src/TowerDefense/modele/Terrain.java
	src/TowerDefense/modele/ennemis/ScorpionSpeciale.java

+++ src/TowerDefense/modele/Tourelle.java
@@ -4,10 +4,7 @@ import TowerDefense.modele.Tours;
 
 public class Tourelle extends Tours {
 
-	
-	
-
-	public Tourelle(double x, double y, Terrain terrain, Jeu jeu) {
+	public Tourelle(int x, int y, Terrain terrain, Jeu jeu) {
 		super(x, y, terrain,jeu);
 	}
