+++ src/TowerDefense/modele/TourelleRoche.java
@@ -8,7 +8,7 @@ import TowerDefense.modele.tourelle.Tourelle;
 
 public class TourelleRoche extends Tourelle {
 
-	public TourelleRoche(double x, double y, Terrain terrain,Jeu jeu) {
+	public TourelleRoche(int x, int y, Terrain terrain,Jeu jeu) {
 		super(x, y, terrain, jeu);
 	}
