+++ src/TowerDefense/modele/tourelle/TourelleTirMultiple.java
@@ -7,7 +7,7 @@ import TowerDefense.modele.tourelle.Tourelle;
 
 public class TourelleTirMultiple extends Tourelle {
 
-	public TourelleTirMultiple(double x, double y, Terrain terrain, Jeu jeu) {
+	public TourelleTirMultiple(int x, int y, Terrain terrain, Jeu jeu) {
 		super(x, y, terrain, jeu);
 	}
 	
@@ -16,14 +16,10 @@ public class TourelleTirMultiple extends Tourelle {
 	//si ce n'est pas le cas alors elle lui lance un projectile
 	//sinon regarde la même chose pour le prochain ennemi de la liste
 	public void tirMultiple() {
-		
 		for(int i=0; i<getJeu().listeEnnemis().size(); i++) {
 			if(ennemisProcheTirTourelleMultiple(50,getJeu().listeEnnemis().get(i)) && !getJeu().projectileExisteSurEnnemi(getJeu().listeEnnemis().get(i).getId())) {
 				getJeu().getListeProjectile().add(new ProjectileRoche(this,getJeu().listeEnnemis().get(i),getJeu()));
 			}
-		}
-		
-		
+		}	
 	}
-
 }
